hashtable: resizing
by default, hash tables will resize. you can invoke forbid_resize() to
prevent this

also, MemMap accepts another param whether it should fault in all the
memory. we now allocate tables with an enormous virtual length, but
only fault the subregion we use for buckets. when we grow, we do NOT
invoke mmap (as it does not scale) but simply increase the 'end' of
the table and relocate items further