epoch: BUG don't unregister slot when thread dies
Not sure of a solution to this, currently. Originally, threads would
die and the slots would not be quiesced, so I created EpochSlotHold
to (1) quiesce the slot and (2) release the slot for future use. The
former fixed the older bug of compaction getting stuck due to the
epoch not progressing, but the latter is a Rust issue, as TLS
destructors are run in unknown order, causing aborts when there are
dependencies among TLS.