postmark: added C++ STL queue
When configured with many transactions and many files, postmark
performance degenerates, spending most time in find_free_file,
linearly scanning the file metadata table it uses for a free slot. I
added a queue to hold the slots which are available for use, appending
new ones and removing from the head.