segment: bug fix in Segment::extend
Each block has a copy of a slice pointing to each block in the
segment. It is used to locate forward blocks when working with objects
larger then BLOCK_SIZE.

When extending a segment with additional blocks, we mistakenly did not
update the back reference (the slice) to include the additional
blocks. The symptom was, in `get_ref`, the number of blocks holding
the object would exceed the amount remaining in the segment, as
indicated by this back reference. There was always one extra block
apparently missing. Turns out, when compaction relocates objects, it
sometimes needs to extend a new segment, and objects placed partially
across the tail blocks would not be retrievable.

Also converted the Segment::head, front, curblk to be plain `usize`
without use of `Option`. These members were never `None`, so this was
unnecessary.

This damn bug took me two days to identify!