segment: redesigned access to freepool
Freepool now a simple Vec guarded by a Mutex. Application threads will
first aquire an MCS lock to avoid stampeding the freepool when rolling
head segments. Furthermore, we create a copy of the freepool size that
the head MCS thread will spin on, to additionally avoid contention on
the freepool lock.

Compaction threads are then free to lock the freepool to release
blocks into.

We create separate allocation paths; one uses the MCS path, the other
directly acquires the freepool mutex (suffixed with a 'p').